pyproject.toml
[project]
name = "bw-proxy"
version = "3.2.0"
description = "A sovereign proxy acting as a security-hardened blind hub between LLMs and the Bitwarden CLI. Zero Trust · AI-Blind · ACID."
readme = "README.md"
authors = [
    { name = "Ivann KAMDEM", email = "kapoivha@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "mcp>=1.26.0",
    "pydantic>=2.12.5",
    "pyyaml>=6.0.3",
    "rich>=14.3.3",
    "typer>=0.12.0",
    "cryptography>=44.0.0",
]

[project.scripts]
bw-proxy = "bw_proxy.main:main"

[build-system]
requires = ["uv_build>=0.10.3,<0.11.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
